// types/quizQuestion/enums.ts
export enum QuizQuestionMessage {
  SUCCESS_CREATE = "Tạo câu hỏi thành công",
  FAIL_CREATE = "Tạo câu hỏi thất bại",
  
  SUCCESS_UPDATE = "Cập nhật câu hỏi thành công",
  FAIL_UPDATE = "Cập nhật câu hỏi thất bại",
  
  SUCCESS_DELETE = "Xóa câu hỏi thành công",
  FAIL_DELETE = "Xóa câu hỏi thất bại",
  
  SUCCESS_GET = "Lấy thông tin câu hỏi thành công",
  FAIL_GET = "Lấy thông tin câu hỏi thất bại",
  
  QUESTION_NOT_FOUND = "Không tìm thấy câu hỏi",
  SUCCESS_REORDER = "Sắp xếp lại câu hỏi thành công",
  FAIL_REORDER = "Sắp xếp lại câu hỏi thất bại",
  UNAUTHORIZED_ACCESS = "Bạn không có quyền truy cập câu hỏi này"
}